Move ansi support to ansi module + add message for unavailable features
