+++ riseml/stream.py
@@ -8,7 +8,7 @@ import stringcase
 
 from riseml.errors import handle_error
 from riseml.client_config import get_stream_url
-
+from riseml.ansi import COLOR_CODES, color_string
 from . import util
 
 ANSI_ESCAPE_REGEX = re.compile(r'\x1b\[(\d+)m')
@@ -21,7 +21,7 @@ class LogPrinter(object):
         self.stream_meta = stream_meta or {}
 
         self.job_ids_color = {
-            id: list(util.COLOR_CODES.keys())[(i + 1) % len(util.COLOR_CODES)]
+            id: list(COLOR_CODES.keys())[(i + 1) % len(COLOR_CODES)]
             for i, id in enumerate(list(self.ids_to_name.keys()))
         }
 
@@ -41,7 +41,7 @@ class LogPrinter(object):
     def _on_message(self, _, message):
         msg = json.loads(message)
         msg_type = msg['type']
-        
+
         if msg_type == 'log':
             self.print_log_message(msg)
         elif msg_type == 'state':
@@ -71,7 +71,7 @@ class LogPrinter(object):
         color = self.job_ids_color[msg['job_id']]
         prefix = "{}| ".format(job_name.ljust(self.indentation))
 
-        return util.color_string(prefix, color=color)
+        return color_string(prefix, color=color)
 
     def print_log_message(self, msg):
         if msg['job_id'] not in self.ids_to_name:
@@ -87,7 +87,7 @@ class LogPrinter(object):
         for partial_line in line.split('\r'):
             last_color = self.job_ids_last_color_used.get(msg['job_id'], 0)
             line_text = "[%s] %s" % (util.str_timestamp(msg['time']),
-                                    util.color_string(partial_line, ansi_code=last_color))
+                                    color_string(partial_line, ansi_code=last_color))
 
             output = "%s%s" % (self._message_prefix(msg), line_text)
             sys.stdout.write(output)
@@ -104,15 +104,15 @@ class LogPrinter(object):
         time = util.str_timestamp(msg['time'])
         state = "[%s] --> %s" % (time, msg['state'])
         output = "%s%s" % (self._message_prefix(msg),
-                           util.color_string(state, color="bold_white"))
+                           color_string(state, color="bold_white"))
         print(output)
         for key in ['reason', 'message', 'exit_code']:
             if msg.get(key, None) is not None:
                 message = "[{}] {}: {}".format(time, stringcase.titlecase(key), msg[key])
-                print("{}{}".format(self._message_prefix(msg), util.color_string(message, color="bold_white")))
+                print("{}{}".format(self._message_prefix(msg), color_string(message, color="bold_white")))
 
     def print_error_message(self, msg):
-        print(util.color_string("Error: {}".format(msg['error']), color="red"))
+        print(color_string("Error: {}".format(msg['error']), color="red"))
 
 def stream_job_log(job):
     ids_to_name = {job.id: job.short_id}
